changed name of main
