+++ touch_button/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS "touch_button_example_main.c"
+idf_component_register(SRCS "touch_button.c"
                     INCLUDE_DIRS "."
                     )
